requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
transformers==4.39.3
torch==2.2.1
pillow==10.3.0
requests==2.31.0
python-multipart==0.0.9
numpy<2
opencv-contrib-python==4.9.0.80
